+++ Vostok.Metrics.System.Tests/CurrentProcessMonitor_Tests.cs
@@ -1,0 +1,18 @@
+﻿using FluentAssertions;
+using NUnit.Framework;
+using Vostok.Metrics.System.Process;
+
+namespace Vostok.Metrics.System.Tests;
+
+[TestFixture]
+internal class CurrentProcessMonitor_Tests
+{
+    [Test]
+    public void CurrentProcessMonitor_ShouldNotThrow_OnMultipleDisposeCalls()
+    {
+        var monitor = new CurrentProcessMonitor();
+        monitor.Dispose();
+        var dispose = () => monitor.Dispose();
+        dispose.Should().NotThrow();
+    }
+}