+++ Vostok.Metrics.System.Tests/HostMetricsCollector_Tests.cs
@@ -23,6 +23,12 @@ namespace Vostok.Metrics.System.Tests
             collector = new HostMetricsCollector();
             collector.Collect();
         }
+        
+        [TearDown]
+        public void TearDown()
+        {
+            collector.Dispose();
+        }
 
         [Test]
         public void Should_measure_cpu_utilization()