+++ Vostok.Metrics.System/Helpers/ReusableFileReader.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using Vostok.Commons.Environment;
 
 namespace Vostok.Metrics.System.Helpers
 {
@@ -33,7 +34,11 @@ namespace Vostok.Metrics.System.Helpers
         private void Reset()
         {
             if (reader == null)
-                reader = new StreamReader(new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite | FileShare.Delete));
+            {
+                // note (kungurtsev, 23.11.2021): buffer size = 1 for disabling FileStream buffering strategy
+                reader = new StreamReader(new FileStream(path, FileMode.Open, FileAccess.Read, FileShare.ReadWrite | FileShare.Delete, 1));
+                return;
+            }
 
             reader.BaseStream.Seek(0, SeekOrigin.Begin);
             reader.DiscardBufferedData();