+++ Vostok.Metrics.System/Host/HostMonitor.cs
@@ -1,6 +1,7 @@
 ﻿using System;
-using System.Collections.Concurrent;
+using System.Collections.Generic;
 using JetBrains.Annotations;
+using Vostok.Commons.Collections;
 using Vostok.Metrics.System.Helpers;
 
 namespace Vostok.Metrics.System.Host
@@ -9,20 +10,52 @@ namespace Vostok.Metrics.System.Host
     /// <see cref="HostMonitor"/> provides means to observe <see cref="HostMetrics"/> periodically.
     /// </summary>
     [PublicAPI]
-    public class HostMonitor
+    public class HostMonitor : IDisposable
     {
-        private readonly ConcurrentDictionary<TimeSpan, PeriodicObservable<HostMetrics>> observables
-            = new ConcurrentDictionary<TimeSpan, PeriodicObservable<HostMetrics>>();
+        private readonly object guard = new object();
 
         private readonly HostMetricsSettings settings;
+
+        private Dictionary<TimeSpan, DisposablePeriodicObservable<HostMetrics>> observables
+            = new Dictionary<TimeSpan, DisposablePeriodicObservable<HostMetrics>>();
 
         public HostMonitor(HostMetricsSettings settings)
             => this.settings = settings;
 
         public HostMonitor()
-            : this (new HostMetricsSettings()) { }
+            : this(new HostMetricsSettings())
+        {
+        }
 
         public IObservable<HostMetrics> ObserveMetrics(TimeSpan period)
-            => observables.GetOrAdd(period, p => new PeriodicObservable<HostMetrics>(p, new HostMetricsCollector(settings).Collect));
+        {
+            lock (guard)
+            {
+                if (observables == null)
+                    throw new ObjectDisposedException(nameof(HostMonitor));
+
+                return observables.GetOrAdd(period, BuildObservable);
+            }
+        }
+
+        public void Dispose()
+        {
+            lock (guard)
+            {
+                if (observables == null)
+                    return;
+                
+                foreach (var disposablePeriodicObservable in observables)
+                    disposablePeriodicObservable.Value.Dispose();
+
+                observables = null;
+            }
+        }
+
+        private DisposablePeriodicObservable<HostMetrics> BuildObservable(TimeSpan period)
+        {
+            var collector = new HostMetricsCollector(settings);
+            return new DisposablePeriodicObservable<HostMetrics>(period, collector.Collect, collector.Dispose);
+        }
     }
 }